LayoutTests/platform/mac-ventura/media/video-source-type-params-expected.txt

EVENT(loadstart)
EXPECTED (stripExtension(relativeURL(video.currentSrc)) == 'content/test') OK
END OF TEST
